refactor: update all module pathes
