+++ manager/__init__.py
@@ -1,0 +1,7 @@
+from .task import TaskManager
+from .vehicle import VehicleManager
+
+__all__ = [
+    "TaskManager",
+    "VehicleManager",
+]
