+++ route/route.py
@@ -1,7 +1,6 @@
 import networkx as nx
 
-from entity.Location import Location
-from entity.Path import Path
+from entity import Location, Path
 
 from route.map import get_rectangle_graph, get_grid_graph, get_yeouido_graph, get_seoul_gu_graph
