+++ tests/manager/test_manage_vehicle.py
@@ -1,5 +1,5 @@
-from entity.Vehicle import Vehicle
-from entity.VehicleManager import VehicleManager
+from entity import Vehicle
+from manager import VehicleManager
 
 
 def test_add_one_vehicle():
@@ -8,8 +8,7 @@ def test_add_one_vehicle():
 
     assert vehicle_manager.get_vehicle("V1").name == "V1"
     assert vehicle_manager.get_vehicle("V1").loc.x == 0
-    assert vehicle_manager.get_vehicle("V1").loc.y == \
-        0
+    assert vehicle_manager.get_vehicle("V1").loc.y == 0
     assert vehicle_manager.get_vehicle("V1").battery == 1
     assert vehicle_manager.get_vehicle("V1").status == Vehicle.WAIT
     assert vehicle_manager.get_vehicle("V1").route == []
